Clean up folder stucture
