+++ src/bin/index.ts
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-import '../lib/bootstrap';
+import '../lib/bin/bootstrap';
 
 import minimist from 'minimist';
 import pjson from 'pjson';
@@ -10,24 +10,26 @@ import { CommandResult } from 'commands';
 import { TakeoffCmdParameters, TakeoffCommandRequest } from 'takeoff';
 import { ExitCode } from 'task';
 
-import checkDependencies from '../lib/check-dependencies';
-import exitWithMessage from '../lib/commands/exit-with-message';
-import fileExists from '../lib/commands/file-exists';
-import getCommandFromString from '../lib/commands/get-command-from-string';
-import getProjectDetails from '../lib/commands/get-project-details';
-import pathExists from '../lib/commands/path-exists';
-import printMessage from '../lib/commands/print-message';
-import createRunCommand from '../lib/commands/run-command';
-import extractArguments from '../lib/extract-arguments';
+// Dependencies for CLI
+import checkDependencies from '../lib/bin/check-dependencies';
+import extractArguments from '../lib/bin/extract-arguments';
+import loadCommands from '../lib/bin/load-commands';
+import loadRcFile from '../lib/bin/load-rc-file';
+
+// Depndencies for Tasks
 import renderHelp from '../lib/help/render-help';
-import loadCommands from '../lib/load-commands';
-import loadRcFile from '../lib/load-rc-file';
+import exitWithMessage from '../lib/helpers/exit-with-message';
+import fileExists from '../lib/helpers/file-exists';
+import getCommandFromString from '../lib/helpers/get-command-from-string';
+import getProjectDetails from '../lib/helpers/get-project-details';
+import pathExists from '../lib/helpers/path-exists';
+import printMessage from '../lib/helpers/print-message';
+import createRunCommand from '../lib/helpers/run-command';
 
 /**
  * Main function executed when the Takeoff commannd line is run
  */
-const run = async (workingDir: string, cliArgs: string[]) => {
-
+async function run(workingDir: string, cliArgs: string[]): Promise<void> {
   const rcFile = loadRcFile(workingDir);
   let customDependencies = [];
   if (rcFile.exists && rcFile.properties.has('dependencies')) {
@@ -64,7 +66,7 @@ const run = async (workingDir: string, cliArgs: string[]) => {
   try {
     takeoffCommands = await loadCommands([`${__dirname}/../commands`, `${workingDir}/commands`], takeoff);
   } catch (e) {
-    throw exitWithMessage({code: ExitCode.Error, fail: 'Unable to load commands', extra: e});
+    throw exitWithMessage({ code: ExitCode.Error, fail: 'Unable to load commands', extra: e });
   }
 
   // Parse the request
@@ -86,18 +88,23 @@ const run = async (workingDir: string, cliArgs: string[]) => {
     });
   }
 
-  // Run the command and exit
-  let result: CommandResult;
+  // Set a default exit, and run the plugin to get the correct result
+  let result: CommandResult = {
+    code: 0,
+    success: `${plugin.group}:${plugin.command} done`,
+  };
   try {
     result = await plugin.handler();
   } catch (e) {
     result = {
       code: ExitCode.Error,
-      fail: e,
+      extra: e,
+      fail: `Unable to run handler for plugin ${plugin.group}:${plugin.command}`,
     };
   }
 
   return exitWithMessage(result);
-};
+}
 
+// Run the CLI tool
 run(process.cwd(), process.argv.slice(2));
