+++ src/index.ts
@@ -1,5 +1,5 @@
-import exitWithMessage from './lib/commands/exit-with-message';
-import printMessages from './lib/commands/print-message';
 import renderTable from './lib/help/render-table';
+import exitWithMessage from './lib/helpers/exit-with-message';
+import printMessages from './lib/helpers/print-message';
 
 export { exitWithMessage, printMessages, renderTable };
