+++ src/lib/bin/check-dependencies.ts
