+++ src/lib/bin/extract-arguments.ts
