+++ src/lib/bin/load-commands.ts
