+++ src/lib/helpers/exit-with-message.ts
