+++ src/lib/helpers/generate-table.ts
