+++ src/lib/helpers/get-command-from-string.ts
