+++ src/lib/helpers/get-project-details.ts
