+++ src/lib/helpers/path-exists.ts
