+++ src/lib/init-env/read-takeoff-file.ts
@@ -1,15 +1,11 @@
 import JoyCon from 'joycon';
-import { dirname } from 'path';
 import { TakeoffFileData } from 'takeoff';
 
 import parseMarkdown from './parse-markdown';
 
 export = (cwd: string): TakeoffFileData => {
   const loadTakeoffFile = new JoyCon({
-    // Stop reading at parent dir
-    // i.e. Only read file from process.cwd()
     cwd,
-    stopDir: dirname(process.cwd()),
   });
 
   const { path: filepath, data } = loadTakeoffFile.loadSync(['takeoff.md']);
