+++ tsconfig.json
@@ -15,10 +15,6 @@
       "esnext"
     ],
     "paths": {
-      "@takeoff/takeoff/types/arguments": ["src/types/arguments.d.ts"],
-      "@takeoff/takeoff/types/tables": ["src/types/tables.d.ts"],
-      "@takeoff/takeoff/types/task": ["src/types/task.d.ts"],
-      "@takeoff/takeoff/types/commands": ["src/types/commands.d.ts"],
       "*": ["node_modules/*", "src/types/*"]
     }
   },
